context: stricter syntax for __context__ statement
The expected syntax for the __context__ statement is:
	__context__(<expression>);
or
	__context__(<context>, <expression>);
but originally it was just:
	__context__ <expression>
In other words the parenthesis were not needed and are
still not needed when no context is given.

One problem with the current way to parse these statements is
that very few validation is done. For example, code like:
	__context__;
is silently accepted, as is:
	__context__ a, b;
which is of course not the same as:
	__context__(a,b);
And code like:
	__context__(,1);
results in a confusing error message:
	error: an expression is expected before ')'
	error: Expected ) in expression
	error: got ,

So, given that:
* the kernel has always used the syntax with parenthesis,
* the two arguments form requires the parenthesis and thus
  a function-like syntax
use a more direct, robust and simpl parsing which enforce
the function-like syntax for both forms.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>