context: fix crashes while parsing '__context__;' or '__context__(;'
The expected syntax for the __context__ statement is:
	__context__(<inc/dec value>);
or
	__context__(<context>, <inc/dec value>);

The distinction between the two formats is made by checking if
the expression is a PREOP with '(' as op and with an comma
expression as inner expression.

However, code like:
	__context__;
or
	__context__(;
crashes while trying to test the non-existing expression
(after PREOP or after the comma expression).

Fix this by testing if the expression is non-null before
dereferencing it.

Note: this fix has the merit to directly address the problem
      but doesn't let a diagnostic to be issued for the case
	__context__;
      which is considered as perfectly valid.
      The next patch will take care of this.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>