return_to_param: fix parsing calls
This is to handle code like spi_get_drvdata().  I think that it
used to partially work, but then I broke it in commit 215c7ada006d
("return_to_param: fix a crashing bug").

The problem there was that I was using an expression pointer instead
of a sym pointer for the sm_state.  That caused a use after free.
The fix was to use the sym from the call->fn instead.  But I never
updated the get_state() call so that was still using the expression
pointer so it didn't work.

But even before I broke it, the code to add to combine "&spi->dev"
with "$0->driver_data" was buggy.  These days we have a function to
take a string and convert it to an expression which works for
spi_get_drvdata().  So I've used that instead.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>