add helper for new parsing errors: unexpected()
Most parsing errors are reported during parsing itself but
sometimes some extra conditions must be checked.
One problem with this is that these extra conditions are not
met because they are the consequence of a primary error which
has already been reported. In this case, it this not desirable
to emit a new error message.

To help with this, add a small function which will emit an
error message followed by the current token but only if the
current token does not originated from a previous parsing error.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>