label: avoid multiple definitions
If a label is defined several times, an error is issued about it.
Nevertheless, the label is used as is and once the code is linearized
several BB are created for the same label and this create
inconsistencies. For example, some code will trigger assertion failures
in rewrite_parent_branch().

Avoid the consistencies by ignoring redefined labels.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>