give a position to end-of-input
If an error occurs at the end of the input, for example because
a missing terminating ';' or '}', the error message is like:
	builtin:0:0: error: ...
IOW, the stream name & position is not displayed because the
because the current token is eof_token_entry which has no position.
This can be confusing and for sure doesn't point where the error is.

Fix this by giving to eof_token_entry the end-of-stream position.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>