avoid multiple error message after parsing error
Parsing error triggered by expect() insert bad_token as the
current token. This allows to skip other errors with expect().
So far so good but this token is a fake token which has no
position set, if another parsing error is detected by something
else than expect(), the error message will use the position of
this bad token which will be displayed like:
	builtin:0:0: ...
which is confusing.

Since the concerned error message are secondary ones, the primary
one have been already be repported by expect(), the best is to
simply skip all these secondary error messages.

Do this by creating a new token type (TOKEN_BAD) and use it for
bad_token, then filter error messages based on this token type.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>