ranges: fix == and != comparisons...  (0_0)
Laurent Pinchart pointed out a bug where we have pointers p1 =
s64min-(-1),(1)-s64max and if (p1 == p2) then p2 should also be non-zero
but it's instead s64min-s64max.

The problem is that == and != were only trimming away the highest and
lowest values.  We should be taking the intersection of p1 and p2.

This is a very old bug and it's sort of shocking that it survived for so
long.

Reported-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>