ptrlist: when possible use the real type of the list
The macros doing the list walking used to be done, not on the declared
list's type but on the concrete & generic type 'struct ptr_list'.
This incurs casting between the real and the generic type.

Now that the declared type matches ... these casts are not needed
anymore and the real declared type can be used instead.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>