ptrlist: remove extra ident level
The list walking macros are relatively complex and the code they define
have often 5 or 6 level of nesting.

However, in a few places, a small rewrite of the code can save one
nesting level and this help to make the code more readable.

Make this change and reduce the indentation level.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>