builtin: add predefine()
Builtin macros are defined via the add_pre_buffer() mechanism
which is quite powerful/flexible but for small, simple builtin
macros, it's a bit sad to have to compose a buffer and then
tokenize it when it would be easy to directly define these macros.

Add a new function, predefine(), which allows to directly add
a definition for a simple macro, taking no args and having
a single number or identifier as definition.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>