equiv: type and implication fixes
There are several bugs here.  First if you set a = b; then a would end up being
the same type as b instead of the casted value.

The second problem is the change to the smatch_extra code.  When we check
if the values fit, we're checking a potentially casted/truncated right
value, as in: "char x = (char)my_integer;"  I've added a strip_expr() to
get right of the cast.

The other problem is that we also updated the related states badly.  We
don't want to change their type or their saved implications, we only want
to update the related list to add the new expression.

I deleted the del_equiv() wrapper and updated the code to use del_related().
Then in del_related() my concern was that actually we shouldn't change the
related list, we should clone the list and remove the unwanted relation.

The "if (ptr_list_size((struct ptr_list *)get_dinfo(state)->related) == 1)"
was a weird unnecessary test.  It's not a special case, just handle it
normally, so I deleted that code.

In set_equiv(), I moved some code around as a cleanup.  We need to do a
__set_sm() because we're going to look up again when we update the related
list.  It's not necessarily the fastest thing to update the new sm_state
twice but I didn't want to deal with a special case.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>