comparison: improve handling of "if (++j == limit)"
Say we have:

	if (++j == limit)

And we start with j < limit.  What we want to record is that on the true
path j == limit and the false path j < limit.

How we go about it is a bit convoluted and sort of special case-ish.  We
first detect that we have code in the format "if (++j == limit)" and we
can know from smatch extra that j < limit.  So then we record that as a
state in smatch_comparison.c.  The the modification code says that if we
start with j < limit and do a ++ that means j <= limit.  Then the normal
comparison code splits it into j < limit and j == limit.

The important bit is that we need to know that we care about j vs limit
when we do the ++.  It's hard to know in advance what relationships are
important and saving everything would probably be too time consuming.  This
might be easier if we turned on the --two-passes option.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>