helper: strip out the byte swaps in strip_expr()
Generally I have adopted the theory that it's easier to just ignore byte
swaps.  That way maybe if someone takes a value, byte swaps it, then byte
swaps it back, I'll know what the value is.

It's a hack, and I'm not positive it works but so far it hasn't hurt
anything that I know of.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>