sparsec: simplify & portable use of mktemp
Most manpages for mktemp(1) specify that the 'XXXXXX' part of the
template must be at the end (of course, the GNU's version accept
it anywhere). At least the busybox's version used by Alpine Linux
tries to enforce that and fails in sparse current use of it:
	tmp.XXXXXX.{llvm,o}

OTOH, the llc command can take its input from stdin and so there
is no need to have 2 temp files that need to be distinguished.

Fix the portability problem by feeding llc via stdin, which removes
the need for the first temp file, and using 'tmp.XXXXXX' as template
for the remaining one.

Fixes: a410611d7 "(llvm: fix creation of sparsec's tmp files")
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>