builtin: add support for __builtin_{add,sub,mul}_overflow(), ...
It seems they will be used in the kernel so add them.

Unlike __builtin_uadd_overflow() and friends, these don't take
a fixed type and thus can't simply be declared but need their
own evaluate() method to do the type checking.

Note: of course, no expansion is done on them.
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>