builtin: add support for __has_builtin()
Sparse has support for a subset of GCC's large collection of builtin
functions. As for GCC, it's not easy to know which builtins are
supported in which versions.

clang has a good solution to this problem: it adds the checking macro
__has_builtin(<name>) which evaluates to 1 if <name> is a builtin
function supported by the compiler and 0 otherwise.
It can be used like:
	#if __has_builtin(__builtin_clz)
	#define clz(x)	__builtin_clz(x)
	#else
	...
	#endif

It's possible or probable that GCC will have this soon too:
	https://gcc.gnu.org/bugzilla/show_bug.cgi?id=66970

Add support for this __has_builtin() macro by extending
the evaluation of preprocessor expressions very much like
it is done to support defined().

Note: Some function-like builtin features, like __builtin_offset(), are
      considered as a kind of keyword/operator and processed as such.
      These are *not* considered as builtins by __has_builtin().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>