teach sparse about -m16
The boot part of the kernel on x86-64 has to compile code for
real-mode. This is 32bit code and is specified to GCC with '-m16'.

However, sparse hadn't knwoledge of this flag and thus uses the
default LP64 model for it which, of course, can cause false
warnings because of the mismatched sizes.

Fix this by teaching sparse about the -m16 flag by treating it
the same as -m32.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>