user_data: introduce is_user_data()
For the Spectre stuff we just care if an offset can be controlled by the
user.  In other words, we don't care what that there is a limit check
which restricts it to 0-7 or whatever because the CPU is speculatively
executing instructions as if the condition wasn't there.

So I have introduced the is_user_rl().

The other difference is that Smatch was doing a hack with certain divides
and marking user controlled values as not user controlled.  I have turned
this off when the --spammy option is used.  Also I have introduced a
get_user_rl_spammy() which does this too.

Reported-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>