user_data: silence some sscanf() false positives
There are two things here.  dev_attr names are trusted.  It's pretty common
to read the name, and get a number from the name and use it.

The other thing is more basic which is that smatch was marking %n as user
controlled when it's not the same as reading a %d from the string.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>