user_data: silence some more ntohl(foo[i]); false positives
Say you had code like:

	ntohl(foo[i]);

Then the problem is that Smatch marks everything from ntohl() as user
controlled.  But the "i" offset into the foo[] array does not come from
ntohl() so it shouldn't be marked as user data.  So we want to take the
parent and see if it's inside the macro.

TBH, I forget the exact details with this...  Sorry.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>