cast: keep instruction sizes consistent
The last instruction of linearize_load_gen() ensure that
loading a bitfield of size N results in a object of size N.
Also, we require that the usual binops & unops use the same type
on their operand and result. This means that before anything
can be done on the loaded bitfield it must first be sign or zero-
extended in order to match the other operand's size.

The same situation exists when storing a bitfield but there
the extension isn't done. We can thus have some weird code like:
	trunc.9		%r2 <- (32) %r1
	shl.32		%r3 <- %r2, ...
where a bitfield of size 9 is mixed with a 32 bit shift.

Avoid such mixing of size and always zero extend the bitfield
before storing it (since this was the implicitly desired semantic).
The combination TRUNC + ZEXT can then be optimised later into
a simple masking operation.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>