extra: use PARAM_LIMIT to limit binops
I was trying to use this code to handle kmalloc(x * sizeof(), GFP_KERNEL);

We know if the kmalloc() fails, then the size is capped, but in this case
the size is a binop so we're not setting anything.  But we'd want to cap
x...  Hence this patch.  It occurs to me now as I write this commit message
that we could do this better inside set_extra_nomod().

Also this code doesn't work great because can_integer_overflow() is too
cautious.  It turns out implementing can_integer_overflow() is quite
tricky...

Anyway, I've been running this for a while and I want to make my diff with
the released code smaller.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>