db: introduce the cache_db
The point of this is that say you want to record some information about
a how a type is used in a file, instead of recording over and over you
can record it in the cache_db.  You'd put restrictions against duplicate
rows or whatever.  Then it's written after we finish parsing the file.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>