use "%Le" to display floats
Floating-point values are displayed using the printf
format "%Lf" but this is the format without exponent
(and with default precision of 6 digit).

However, by its nature, this format is very imprecise.
For example, *all* values smaller than 0.5e-6 are displayed
as "0.000000".

Improve this by using the "%Le" format which always use
an exponent and thus maximize the precision.

Note: ultimately, we should display them exactly, for
      example by using "%La", but this will requires C99.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>