db: rename call_implies to return_implies
I have caller_info and return_states which are the states going into and
out of functions.  The call_implies table is sort of like return_states
but it applies to every path...  Mostly it's used for parsing the returned
states to say that "this parameter foo->bar->baz was used" or "we
dereferenced that parameter".  So really it should be called return_implies
instead of call_implies.

I have sort of started to add some call_implies states to the
return_implies table (although maybe none of that work was published) but
it's the wrong thing to mix the two different sorts of states.  So now I'm
going to introduce a new call_implies table and rename the old one to
return_implies.

I've created the new call_implies table but not put the plumbing into place
to make it useful.  The reason I did that in this commit is to avoid
annoying error messages.  It won't work either way until you rebuild the
DB, but it's slightly less annoying if you don't get a ton of error
messages.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>