user_data: don't pass *$ as user data if the struct members are known
Say you have a void pointer or a char pointer.  Then we record that the
data it points to is user data.  But if it's a struct member then we set
all the struct members so we don't need the *$ information any more.  And
that information can be misleading if the struct members are reset to
non-user data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>