nospec: if it's called with trusted data that's the same as nospec
The problem here is that when we mix user, nospec data with trusted data,
the result should be nospec.

I made this only true when we're dealing with caller_info.  It should
probably just be true all the time.  I think it doesn't matter and this
felt cautious and conservative to me when I wrote it.

Reported-by: Josh Poimboeuf <jpoimboe@redhat.com>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>