cast: simplify [SZ]EXT + TRUNC to a smaller/greater size
An OP_SEXT or a OP_ZEXT followed by a truncate to a size smaller
than the original size is unneeded, the same result can be obtained
by doing the truncate directly on the original value.

Dualy, an OP_SEXT or a OP_ZEXT followed by a truncate to a size greater
than the original size doesn't need the truncate, the same result can be
obtained by doing the extend directly on the original value.

Rearrange the inputs (src & orig_type) to bypass the unneeded operation.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>