cast: simplify [ZS]EXT(AND(x,M),N)
A OP_AND with a constant value followed by a sign or zero-extension
can often be moved after the extension.

This is good because it can trigger the cancellation of the OP[ZS]EXT
with a preceding OP_TRUNC, a common situation with bitfields.

Note: This 'simplification' is less desirable if there is
      no preceding OP_TRUNC, for example, in a sequence like:
		and.32	%t <- %a, $mask
		*ext.64 %r <- %t
      If needed, this can be filtered out at some later stage.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>