check_user_data: mark kvm_register_read() as returning user data
The main thing from this patch is:

add_function_assign_hook("kvm_register_read", &match_returns_user_rl, NULL);

But I wanted to rename the function because that's a simple_strtoul
function.  Then I noticed that we should be taking the type of the function
and casting that range list to the variable we're assigning it to.  It
probably doesn't make a big difference in real life, but it's more correct.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>