add lookup_ptr_list_entry()
For liveness analysis, the ptrlists need to be used as sets. IOW,
before adding a new element, it's needed to check if it doesn't
already belong to the list.

This check is currently done, specifically for pseudos,  using the
list walking macros FOR_EACH_PTR/END_FOR_EACH_PTR.

Add a new generic ptrlist function: lookup_ptr_list_entry() which
test if a given pointer already belong to the list.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>