add copy_ptr_list()
When an instruction can be replaced by a pseudo, the user list of
this pseudo and the instruction's target must be merged.

Currently this is done by concat_ptr_list() which copy the elements
of one list into the other using add_ptr_list(). This incurs quite
a bit overhead.

Add a new more efficient ptrlist function: copy_ptr_list() which
copy the element by block by looping over both list in parallel.

This gives a speedup up to 26% on some pathological workloads.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>