add ptr_list_empty()
Sometimes we need to know if a list is empty, for example, in
order to determine if a pseudo has some users or not.

Currently, this is done using ptr_list_size(), which always
walks the whole list but the needed answer can be returned as
soon as it's known that the list contains at least one element.

Add the helper ptr_list_empty() and use it for has_users().

This gives a speedup up to 18% on some pathological workloads.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>