add ptr_list_multiple()
When doing IR simplification, to know if an instruction can be
destructively modified, it's needed to know if the pseudo it defines
is used by a single instruction or not. Currently this is done using
ptr_list_size() which needs to walk the whole list.

This walk is relatively costly when the list is long but knowing
if the list contains more than 1 element can often be answered
more cheaply since an answer can be returned as soon as it's
known that the list contains at least 2 elements.

Add the helpers ptr_list_multiple() and multi_users().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>