spectre: allow to flag all occurrences
The initial version of the scanner flagged each problematic array only
once. This results in warnings being suppressed, or manual code inspection
where the tool could actually help much more. Therefore, we introduce a
flag that allows to switch between this behavior, or printing all possible
warnings. This way, the manual work is to triage more warnings, but not to
search for further candidates.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>