conditions: split the zero
I have a check, which I should publish, which warns about if you do
"UINT_MAX + 1" because that's an integer overflow.  But it missed a bug
where we did "if (x == UINT_MAX + 1) ".  The problem is that the code
in smatch_conditions.c munges away comparisons to zero.

If it's not a EXPR_VALUE, let's split the expression like we normally
would.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>