expand linearize_position() into linearize_initializer()
linearize_position() is a very small function doing the initialization
of EXPR_POS while for the more complex EXPR_INITIALIZER this is done
directly in linearize_initializer().

Having this code in a separate function doesn't bring substantial benefits,
on the contrary it obscure the fact that linearize_initializer() is a
recursive function and make harder the tracking of where struct access_data
are modified.

Expand linearize_position() into its caller, linearize_initializer().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>