put back the bitfield base type into struct access_data
In commit 48bda7573 ("let struct access_data use a single type"),
the members result_type & source_type have been replaced by an unique
member 'type', mainly because the role of these members and when they
needed to be updated was not clear at all.

Now that the situation is much clearer, the second member, used to
hold the base type of bitfields, can be put back as it avoid
to have to call bitfield_base_type() multiple time.

Put back this member into struct access_data and use when appropriate
(which remove the now unneeded calls to bitfield_base_type()).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>