fix instruction size & type in linearize_inc_dec()
If the ++ or -- operator is used on a bitfield, the addition or
subtraction is done with the size of the bitfield. So code like:
	struct {
		int f:3;
	} s;
	...
	s->f++;
will generate intermediate code like:
	add.3	%r <- %a, $1

This is not incorrect from the IR point of view but CPUs have
only register-sized instructions, like 'add.32'. So, these
odd-sized instruction have one or two implicit masking/extend
that should better make explicit.

Fix this by casting to and from the base type when these operators
are used on bitfields.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>