shift: simplify ASR(ZEXT(X, N), C)
After an OP_ZEXT, it is guaranteed that the sign bit is zero.
So, an arithmetic right-shift following an OP_ZEXT gives the
same result as an logical right-shift which is simpler to handle
and further optimize.

So, when preceded by an OP_ZEXT, replace OP_ASR by OP_LSR.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>