big-shift: fix evaluation of shift-assign
The right hand side of a shift operation must be of integer type
and do the integer promotion and nothing else. More precisely, the
RHS and the LFS type doesn't need to match.

This is correctly done for plain shifts but not for shift-assigns
where the full usual conversion is wrongly applied.

Fix this by special casing shifts in evaluate_assign_op() and
just apply the integer promotion on the RHS, like done for shifts
in evaluate_binop().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>