big-shift: fix warning message for negative or over-sized shifts
In order to preserve shifts with negative count and give a proper
warning for them, we first need to identify when the shift must
be considered as negative. This identification is made complicated
by the fact that PSEUDO_VAL are typeless.

The solution chosen here is to try to first sign extend the shift
count (with the size of an int and the size of the instruction).
If after this extension the count appears as negative it is treated
as such (and this new value replace the old one to avoid to have to
redo this at each simplification run), otherwise it is considered
as unsigned. The only case where a negative count could be wrongly
interpreted as a big unsigned value would be if:
	size target int < instruction's size < size host value
which can't happen for now since sizeof(int) == 4 on all targets
and sizeof(value) is 8.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>