conditions: set the parent expressions
When we're parsing conditions, then a lot of the expressions don't go
through __split_expr().  As a result, the parent pointers weren't getting
set correctly.  Some examples are negatives and parenthesis because we
want to ignore these as much as possible.  These are all removed in
strip_expr() so I made a function called strip_expr_set_parent() which
sets the parent pointers.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>