db: Use extra sm->possible to decide if splitting the returns makes sense
If there isn't a separate NULL or error pointer state then it doesn't
make sense to try to split the states apart.  There are two problems with
splitting it apart in that case.  You end up thinking that the function
can return NULL when really it just returns unknown.  And also you end
storing duplicate data in the DB.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>