db: hook up call_implies infrastructure
call_implies and return_implies are very similar and we can re-use some
of the code.  return_implies means that say "every path through the
function dereferenced a pointer".  call_implies means that at the start
of the function we know that param 2 is the size of the param 1 buffer.

But one trick is that not only is call_implies called at the start of the
function call, it's also hooked into the caller so the caller can use
the information that param 2 is the size of the buffer.

I generally think we will want to push all the call_implies iformation into
the cache_db before printing it.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>