type: introduce get_promoted_type()
One common thing is that we want to parse a binop "a + b" and we want to
get the type to promote it to.  I've created a function for this.

Also if you add an unsigned long to a pointer the result is still a
pointer, so in this sense this is a bug fix.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>