address: fix add_offset_to_pointer()
The problem is I was taking an mtag and adding an offset like so
"&foo->bar" where foo has an mtag and this function was returning an
empty rl.

Then the other problem was that I changed rl_binop() to take alignment
into consideration.  In this function we're not doing pointer math, we are
just adding 16 bytes to &foo.

Instead of figuring out the exact issue, I just re-wrote the function.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>