container_of: flip the param and the value around
Before how this worked was we said that -1 is related to param but I really
want to say that param is related to -1.  The reason is that you could have
two parameters which are related to the function pointer.  It doesn't
really matter either way, really.  But this way is slightly less code as
well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>