cast: fix warning position in cast_pseudo()
The function cast_pseudo() can issues warnings about non size-preserving
integer <-> pointer casts. The file:line:column position of these warnings
is taken from the destination type but this position is the one where the
type is declared (or where the symbol associated with this type is defined)
which may or may not be related to the position of the cast.

Fix this by using the current position instead (which should hold the
position of the cast).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>