container_of/mtag: fix handling of passed values (as opposed to pointers)
There are actually a bunch of changes in this code.  The main bug fix is
in load_tag_info_sym() where it used to insist that the argument was a
pointer otherwise it would just do nothing.

The rest is parts of re-writes I'm working on.  I want to save things in a
stree and eventually get rid of set_param_value(), and also I changed the
arg_offset stuff from being negative.  I'm not positive the call:

	if (!mtag_map_select_tag(container_tag, -arg_offset, &arg_tag))

is correct or useful.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>