mtag: turn off mtag stuff for 32 bit builds
I'm not really comfortable with this, but for now I don't want to think
about 32 bit systems.  The issue is that I'm mtags are a hash and ideally
they are unique.  I use 1 bit for if it's an alias or a base mtag, and 12
bits for the offset and on 32 bit systems that would leave only 19 bits
for the hash.  It feels like we would get all kinds of collisions.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>