mtag/math/extra: start using mtags as pointers
Say you have code like:

	struct foo *p = some_array;

Then really we know that some_array is non-NULL, but we don't know where
the linker will put it.  In other words we don't know the actual address.
But let's use mtags and *pretend* that we do know where it's going to be.

So now in smatch extra will record that p is a specific value instead of
just saying that it's non-NULL.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>