mtag: make get_mtag_offset() give both an mtag and an offset
From the name, you'd sort of expect it to return an mtag and an offset.
Also there is a kind of complicated thing with pointers where it hard to
talk about the pointer vs what the pointer is holding.  What's the value
of a pointer?  Is it the address or the thing stored in the address.

This function is used when you want to get the thing stored in the address.
But really it gives you the address...  I don't know if this explanation
is helping...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>