mtag/math: use mtag information
This just hooks the mtag infrastructure in so that it can be used.

I'm not totally sure what state the mtag code is in.  I wrote it some time
back and I have been using it but it could be that it doesn't actually
do anything.  Or it could slightly buggy, and I have forgotten.  Anyway,
it makes testing easier if I push this change.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>