mtag: improve get_mtag_sval()
Say we want to find out "__smatch_implied(&array);" then get_mtag_sval()
is the function to use.

One change that I made is that for the address of array elements like
"__smatch_implied(&array[4]);" then originally I had thought that I should
create a new mtag for each element, but now I think that it should be just
the mtag of the array plus the offset.  The mtag sort of like a symbol in
sparse/smatch and then we add an offset to that to get the actual address.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>