mtag: Use extra state to get mtag and offset
We need to take a look at validation/sm_mtag6.c to really explain what
this code is doing.  We call:

	frob2(&aaa.b);

The "aaa" struct is a top level struct.  We know what the mtag sval is for
&aaa.b so we pass that to frob2().  If we built the DB twice then we
would also know the value of "aaa.b" and we would pass that but for this
test we're only passing it once.

So then when we parse frob2() with the DB we know the "value" (meaning
the mtag sval) of "p".  So we look it up and get the value of *p.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>